testing/web-platform/tests/html/syntax/parsing-html-fragments/the-input-byte-stream-037.html

<!DOCTYPE html>
<html  lang="en" >
<head>
 <meta http-equiv="content-type" content="text/html; charset=iso-8859-15"> <title>UTF-8 BOM vs meta content</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/html5/syntax.html#the-input-byte-stream'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='http'>
<style type='text/css'>
.test div { width: 50px; }</style>
<link rel="stylesheet" type="text/css" href="support/encodingtests-utf8.css">
</head>
<body>



<div class='test'><div id='box' class='ýäè'>&#xA0;</div></div>









<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 100);
}, "A page with a UTF-8 BOM will be recognized as UTF-8 even if the meta content attribute declares a different encoding.");
</script>

<div id='log'></div>

</body>
</html>
